requirements.txt
colorama>=0.4.0
psutil>=5.8.0
python-dotenv>=0.19.0
fastmcp>=2.0.0
strands-agents>=1.9.1
strands-agents-tools>=0.2.8
mcp>=1.0.0
openapi-spec-validator>=0.5.0
pydantic>=1.10.0
requests>=2.25.0
httpx>=0.25.0